Add modules dependency
Added a module dependency. From IntelliJ docs:

"If a plugin does not include any module dependency tags in its plugin.xml,
it’s assumed to be a legacy plugin and is loaded only in IntelliJ IDEA."

By adding a module dependency, Swagger plugin works with PyCharm and RubyMine
for example.