Use default editor in addition to Swagger UI
Swagger UI split view does not support all the features that
the intellij default editor has. One example is the structure view
that is missing with the Swagger UI split view.

This commit enables the use of the default editor, so that users
have a choice to use either one.