Fix Swagger UI
The new version of `gradle-intellij-plugin` packages the
plugin in a different way. From now on, the plugin is packaged
to a single intellij-swagger.jar file. In order for the swagger UI
file copying to work, the code needed to be
updated (we can't use absolute file paths anymore, as we are inside
a jar file).

Fixes #230