Improve $ref validation
The previous setup was hard to maintain with individual
code handling different types of references and looking for
the referenced elements manually.

This commit simplifies the $ref validation by using visitors
and reporting errors for any $ref values that resolve to zero elements.

The implementation also adds validation for file paths that
was missing before.

Part of #226